+++ packages/react-docsite-components/src/components/Page/Page.types.ts
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { IComponentAs } from '@fluentui/react';
 import { IExampleCardProps } from '../ExampleCard/index';
 import { ISideRailLink } from '../SideRail/index';
-import { IPageJson } from '@fluentui/react-internal/lib/common/DocPage.types';
+import { IPageJson } from '@fluentui/react/lib/common/DocPage.types';
 
 /**
  * Props for the page.