+++ packages/react-examples/src/azure-themes/stories/Themes/Themes.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Customizer, TextField, Stack, Fabric, Checkbox, SearchBox, Link, Label, Text } from '@fluentui/react';
-import { DefaultButton, CompoundButton, PrimaryButton } from '@fluentui/react/lib/compat/Button';
+import { DefaultButton, CompoundButton, PrimaryButton } from '@fluentui/react/lib/Button';
 import {
   AzureCustomizationsLight,
   AzureCustomizationsDark,