+++ packages/react-examples/src/react/Breadcrumb/Breadcrumb.doc.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { BreadcrumbBasicExample } from './Breadcrumb.Basic.Example';
 import { BreadcrumbCollapsingExample } from './Breadcrumb.Collapsing.Example';
 import { BreadcrumbStaticExample } from './Breadcrumb.Static.Example';
-import { IDocPageProps } from '@fluentui/react-internal/lib/common/DocPage.types';
+import { IDocPageProps } from '@fluentui/react/lib/common/DocPage.types';
 
 const BreadcrumbBasicExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/Breadcrumb/Breadcrumb.Basic.Example.tsx') as string;
 const BreadcrumbCollapsingExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/Breadcrumb/Breadcrumb.Collapsing.Example.tsx') as string;