+++ packages/react-icons-mdl2/src/utils/createSvgIcon.ts
@@ -3,10 +3,12 @@ import { css, getNativeProps, htmlElementProperties } from '@fluentui/utilities'
 import * as classes from './SvgIcon.scss';
 import { ISvgIconProps } from './SvgIcon.types';
 import { SvgIconCreateFnParams } from './types';
+import { useIconSubset } from '@fluentui/react-icon-provider';
 
 const createSvgIcon = <TProps = {}>({ svg, displayName }: SvgIconCreateFnParams<TProps>) => {
   const Component: React.FC<React.HTMLAttributes<HTMLSpanElement> & TProps & ISvgIconProps> = props => {
     const { className, style = {} } = props;
+    const icons = useIconSubset(); // TODO: handle fontFace and styles
 
     const nativeProps = getNativeProps<React.HTMLAttributes<HTMLElement>>(props, htmlElementProperties);
     const containerProps = props['aria-label']
@@ -15,7 +17,6 @@ const createSvgIcon = <TProps = {}>({ svg, displayName }: SvgIconCreateFnParams<
           role: 'presentation',
           ['aria-hidden']: true,
         };
-
     return React.createElement(
       'span',
       {
@@ -24,7 +25,7 @@ const createSvgIcon = <TProps = {}>({ svg, displayName }: SvgIconCreateFnParams<
         className: css(classes.root, className),
         style,
       },
-      svg({ classes, props }),
+      icons?.icons?.[displayName] ? icons.icons[displayName] : svg({ classes, props }),
     );
   };