+++ apps/public-docsite-resources/CHANGELOG.md
@@ -1,8 +1,43 @@
 # Change Log - @fluentui/public-docsite-resources
 
-This log was last generated on Mon, 08 Feb 2021 12:23:08 GMT and should not be manually modified.
+This log was last generated on Tue, 09 Feb 2021 12:24:19 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.0.5](https://github.com/microsoft/fluentui/tree/@fluentui/public-docsite-resources_v1.0.5)
+
+Tue, 09 Feb 2021 12:24:19 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/public-docsite-resources_v1.0.4..@fluentui/public-docsite-resources_v1.0.5)
+
+### Patches
+
+- Combine react-internal back into react, and update references ([PR #16832](https://github.com/microsoft/fluentui/pull/16832) by elcraig@microsoft.com)
+
+### Changes
+
+- Bump @fluentui/react to v8.0.0-beta.51 ([PR #16832](https://github.com/microsoft/fluentui/pull/16832) by tristan.watanabe@gmail.com)
+- Bump @fluentui/azure-themes to v8.0.0-beta.50 ([PR #16832](https://github.com/microsoft/fluentui/pull/16832) by tristan.watanabe@gmail.com)
+- Bump @fluentui/theme-samples to v8.0.0-beta.50 ([PR #16832](https://github.com/microsoft/fluentui/pull/16832) by tristan.watanabe@gmail.com)
+- Bump @fluentui/react-docsite-components to v8.0.0-beta.51 ([PR #16832](https://github.com/microsoft/fluentui/pull/16832) by elcraig@microsoft.com)
+- Bump @fluentui/react-monaco-editor to v1.0.0-beta.50 ([PR #16832](https://github.com/microsoft/fluentui/pull/16832) by elcraig@microsoft.com)
+
+## [1.0.4](https://github.com/microsoft/fluentui/tree/@fluentui/public-docsite-resources_v1.0.4)
+
+Tue, 09 Feb 2021 00:56:52 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/public-docsite-resources_v1.0.3..@fluentui/public-docsite-resources_v1.0.4)
+
+### Patches
+
+- Bump @fluentui/react-examples to v1.1.0 ([PR #16865](https://github.com/microsoft/fluentui/pull/16865) by lingfan.gao@microsoft.com)
+
+### Changes
+
+- Bump @fluentui/react-internal to v8.0.0-beta.44 ([PR #16835](https://github.com/microsoft/fluentui/pull/16835) by ololubek@microsoft.com)
+- Bump @fluentui/react to v8.0.0-beta.50 ([PR #16835](https://github.com/microsoft/fluentui/pull/16835) by ololubek@microsoft.com)
+- Bump @fluentui/azure-themes to v8.0.0-beta.49 ([PR #16835](https://github.com/microsoft/fluentui/pull/16835) by ololubek@microsoft.com)
+- Bump @fluentui/react-docsite-components to v8.0.0-beta.50 ([PR #16835](https://github.com/microsoft/fluentui/pull/16835) by ololubek@microsoft.com)
+- Bump @fluentui/react-monaco-editor to v1.0.0-beta.49 ([PR #16835](https://github.com/microsoft/fluentui/pull/16835) by ololubek@microsoft.com)
+- Bump @fluentui/theme-samples to v8.0.0-beta.49 ([PR #16835](https://github.com/microsoft/fluentui/pull/16835) by ololubek@microsoft.com)
 
 ## [1.0.3](https://github.com/microsoft/fluentui/tree/@fluentui/public-docsite-resources_v1.0.3)